refactor(admin): improve agency detail components with type safety and code cleanup
Enhance type annotations for UI helpers in AgencyStatClient, refactor AgencyRequestDetail to client-side component with action modals, simplify page structure, and remove unused agency update methods from rating service for better maintainability.
