+++ apps/frontend/src/app/admin/agencies/create/page.tsx
@@ -1,0 +1,8 @@
+"use client"
+export default function Page() {
+    return (
+        <>
+        <h1>Hello world</h1>
+        </>
+    )
+}
