+++ apps/frontend/src/app/admin/agency-request/detail/[id]/page.tsx
@@ -1,249 +1,9 @@
 "use client";
-
-import { useParams, useRouter } from "next/navigation";
-import { useAgencyActions } from "@/hooks/use-agency";
-import { Button } from "@/components/ui/button";
-import { Badge } from "@/components/ui/badge";
-import {
-  MapPin,
-  Phone,
-  Mail,
-  User,
-  Building2,
-  Warehouse,
-  ArrowLeft,
-  CheckCircle2,
-  XCircle,
-  Calendar,
-  Info,
-  Bike,
-  Navigation,
-  Loader2,
-} from "lucide-react";
-import { cn } from "@/lib/utils";
-import { formatToVNTime } from "@/lib/formatVNDate";
-import ActionRequestModal from "../../components/ActionModal";
-import { useState, useEffect } from "react";
-const DetailSection = ({ icon: Icon, title, children }: any) => (
-  <div className="rounded-xl border border-border/60 bg-card overflow-hidden flex flex-col shadow-sm h-full">
-    <div className="flex items-center gap-2 border-b border-border/60 px-5 py-4 bg-muted/20">
-      <Icon className="h-5 w-5 text-primary" />
-      <h2 className="text-sm font-semibold uppercase tracking-wider">{title}</h2>
-    </div>
-    <div className="p-5 flex-1 space-y-4">{children}</div>
-  </div>
-);
-
-const DetailRow = ({ label, value, icon: Icon }: any) => (
-  <div className="flex flex-col gap-1 border-b border-border/40 pb-3 last:border-0 last:pb-0">
-    <span className="text-xs text-muted-foreground flex items-center gap-1">
-      {Icon && <Icon className="h-3 w-3" />} {label}
-    </span>
-    <span className="text-sm font-medium">{value || "Chưa cập nhật"}</span>
-  </div>
-);
-
-export default function AgencyRequestDetail() {
-  const { id } = useParams();
-  const router = useRouter();
-
-  const [modalState, setModalState] = useState<{
-    isOpen: boolean;
-    type: "APPROVE" | "REJECT" | null;
-  }>({ isOpen: false, type: null });
-  
-  const [isVisualLoading, setIsVisualLoading] = useState(true);
-  const [isProcessing, setIsProcessing] = useState(false);
-
-  const {
-    agencyRequestDetail: data,
-    isLoadingAgencyRequestDetail,
-    approveAgencyRequest,
-    rejectAgencyRequest,
-    getAgencyRequestDetail,
-  } = useAgencyActions({
-    hasToken: true,
-    agency_request_id: id as string,
-  });
-  useEffect(() => {
-    if (id) {
-      getAgencyRequestDetail();
-    }
-  }, [id, getAgencyRequestDetail]);
-
-  useEffect(() => {
-    if (isLoadingAgencyRequestDetail) {
-      setIsVisualLoading(true);
-    } else {
-      const timer = setTimeout(() => {
-        setIsVisualLoading(false);
-      }, 600);
-      return () => clearTimeout(timer);
-    }
-  }, [isLoadingAgencyRequestDetail]);
-
-  const handleConfirmAction = async (payload: {
-    description?: string;
-    reason?: string;
-  }) => {
-    setIsProcessing(true);
-    try {
-      if (modalState.type === "APPROVE") {
-        await approveAgencyRequest({
-          id: id as string,
-          description: payload.description,
-        });
-      } else {
-        await rejectAgencyRequest({
-          id: id as string,
-          reason: payload.reason,
-          description: payload.description,
-        });
-      }
-      setModalState({ isOpen: false, type: null });
-      router.refresh();
-    } finally {
-      setIsProcessing(false);
-    }
-  };
-
-  if (isVisualLoading) {
-    return (
-      <div className="flex h-[70vh] w-full flex-col items-center justify-center gap-4">
-        <Loader2 className="h-10 w-10 animate-spin text-primary" />
-        <p className="text-sm font-medium text-muted-foreground animate-pulse">
-          Đang tải thông tin chi tiết...
-        </p>
-      </div>
-    );
-  }
-
-  if (!data && !isVisualLoading) {
-    return (
-      <div className="flex flex-col items-center justify-center h-[50vh] gap-4">
-        <p className="text-muted-foreground font-medium text-lg">Không tìm thấy yêu cầu này.</p>
-        <Button onClick={() => router.back()}>Quay lại</Button>
-      </div>
-    );
-  }
-
+import AgencyRequestDetailClient from "./AgencyRequestDetail";
+export default function Page() {
   return (
-    <div className="space-y-6 max-w-7xl mx-auto animate-in fade-in duration-500">
-      <div className="flex flex-col gap-4 rounded-xl border border-primary/10 bg-white p-6 sm:flex-row sm:items-center sm:justify-between shadow-sm">
-        <div className="flex items-center gap-4">
-          <Button variant="ghost" size="icon" onClick={() => router.back()} className="rounded-full">
-            <ArrowLeft className="h-5 w-5" />
-          </Button>
-          <div className="space-y-1">
-            <div className="flex items-center gap-3">
-              <h1 className="text-xl font-bold tracking-tight">Chi tiết yêu cầu Agency</h1>
-              <Badge className={cn(
-                  "font-medium",
-                  data?.status === "PENDING"
-                    ? "bg-amber-100 text-amber-700 hover:bg-amber-100"
-                    : data?.status === "APPROVED"
-                      ? "bg-green-100 text-green-700 hover:bg-green-100"
-                      : "bg-red-100 text-red-700 hover:bg-red-100",
-                )}
-              >
-                {data?.status}
-              </Badge>
-            </div>
-            <p className="text-xs text-muted-foreground flex items-center gap-1">
-              <Calendar className="h-3 w-3" /> Gửi ngày: {data?.createdAt ? formatToVNTime(data.createdAt) : "N/A"}
-            </p>
-          </div>
-        </div>
-
-        {data?.status === "PENDING" && (
-          <div className="flex gap-2">
-            <Button
-              variant="outline"
-              className="text-amber-700 border-amber-200 hover:bg-amber-50"
-              onClick={() => setModalState({ isOpen: true, type: "REJECT" })}
-            >
-              <XCircle className="mr-2 h-4 w-4" /> Từ chối
-            </Button>
-            <Button
-              className="bg-green-600 hover:bg-green-700 shadow-md shadow-green-200 text-white font-medium"
-              onClick={() => setModalState({ isOpen: true, type: "APPROVE" })}
-            >
-              <CheckCircle2 className="mr-2 h-4 w-4" /> Phê duyệt
-            </Button>
-          </div>
-        )}
-      </div>
-      <div className="grid gap-6 md:grid-cols-2 lg:grid-cols-3">
-        <DetailSection icon={User} title="Người yêu cầu">
-          <DetailRow label="Họ tên" value={data?.requesterUser?.fullName} icon={User} />
-          <DetailRow label="Email liên hệ" value={data?.requesterEmail} icon={Mail} />
-          <DetailRow label="Số điện thoại" value={data?.requesterPhone} icon={Phone} />
-        </DetailSection>
-
-        <DetailSection icon={Building2} title="Thông tin Agency">
-          <DetailRow label="Tên thương hiệu" value={data?.agencyName} icon={Building2} />
-          <DetailRow label="Địa chỉ văn phòng" value={data?.agencyAddress} icon={MapPin} />
-          <DetailRow label="SĐT kinh doanh" value={data?.agencyContactPhone} icon={Phone} />
-        </DetailSection>
-
-        <DetailSection icon={Warehouse} title="Thông tin Trạm xe">
-          <DetailRow label="Tên trạm" value={data?.stationName} icon={Warehouse} />
-          <DetailRow label="Địa chỉ đặt trạm" value={data?.stationAddress} icon={MapPin} />
-          <div className="grid grid-cols-2 gap-4 pt-2">
-            <div className="p-3 bg-muted/30 rounded-lg">
-              <p className="text-[10px] text-muted-foreground uppercase font-bold">Sức chứa</p>
-              <p className="text-lg font-bold text-primary flex items-center gap-1">
-                <Bike className="h-4 w-4" /> {data?.stationTotalCapacity}
-              </p>
-            </div>
-            <div className="p-3 bg-muted/30 rounded-lg">
-              <p className="text-[10px] text-muted-foreground uppercase font-bold">Lượt Pick/Return</p>
-              <p className="text-sm font-bold">{data?.stationPickupSlotLimit} / {data?.stationReturnSlotLimit}</p>
-            </div>
-          </div>
-        </DetailSection>
-      </div>
-
-      <div className="grid gap-6 lg:grid-cols-2">
-        <DetailSection icon={Navigation} title="Vị trí tọa độ">
-          <div className="p-4 bg-muted/20 rounded-lg border border-dashed border-border flex justify-around">
-            <div className="text-center">
-              <p className="text-xs text-muted-foreground">Vĩ độ (Latitude)</p>
-              <code className="text-sm font-mono font-bold text-blue-600">{data?.stationLatitude}</code>
-            </div>
-            <div className="border-r border-border" />
-            <div className="text-center">
-              <p className="text-xs text-muted-foreground">Kinh độ (Longitude)</p>
-              <code className="text-sm font-mono font-bold text-blue-600">{data?.stationLongitude}</code>
-            </div>
-          </div>
-        </DetailSection>
-
-        <DetailSection icon={Info} title="Thông tin bổ sung">
-          <div className="space-y-4">
-            <div>
-              <div className="text-xs text-muted-foreground uppercase mb-1">Mô tả từ người gửi:</div>
-              <div className="p-3 bg-muted/30 rounded-lg text-sm min-h-[60px]">
-                {data?.description || "Không có mô tả kèm theo."}
-              </div>
-            </div>
-            {data?.status !== "PENDING" && (
-              <div className="p-3 bg-primary/5 border border-primary/10 rounded-lg">
-                <p className="text-xs font-bold text-primary uppercase">Người xử lý:</p>
-                <p className="text-sm font-medium">{data?.reviewedByUser?.fullName || "Hệ thống"}</p>
-              </div>
-            )}
-          </div>
-        </DetailSection>
-      </div>
-
-      <ActionRequestModal
-        isOpen={modalState.isOpen}
-        type={modalState.type}
-        isLoading={isProcessing}
-        onClose={() => setModalState({ isOpen: false, type: null })}
-        onConfirm={handleConfirmAction}
-      />
-    </div>
+    <>
+      <AgencyRequestDetailClient />
+    </>
   );
 }
