+++ apps/frontend/src/services/rating.service.ts
@@ -2,7 +2,6 @@ import fetchHttpClient from "@/lib/httpClient";
 import type { AxiosResponse } from "axios";
 import type { Rating } from "@/types";
 import { ApiResponse } from "@/types";
-import { type UpdateStatusAgencyFormData , type UpdateAgencyFormData, updateStatusAgencySchema } from "@/schemas";
 import { ENDPOINT } from "@/constants";
 export const ratingService = {
   getAllRatings: async ({
@@ -27,18 +26,4 @@ export const ratingService = {
     );
     return response;
   },
-  updateAgency : async ({data,id} : {data : Partial<UpdateAgencyFormData> , id : string}) : Promise<AxiosResponse<Rating>> => {
-    const response = await fetchHttpClient.patch<Rating>(
-      ENDPOINT.RATING.ID(id),
-      data
-    );
-    return response;
-  },
-  updateStatusAgency : async({data,id}:{data : UpdateStatusAgencyFormData , id : string}) : Promise<AxiosResponse<Rating>> => {
-    const response = await fetchHttpClient.patch<Rating>(
-      ENDPOINT.AGENCY.STATUS(id),
-      data
-    );
-    return response;
-  }
 };
